[#38] refactor: Change button location to atoms
