+++ src/components/atoms/button/buttonNormal.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import { TouchableOpacity, TouchableOpacityProps } from "react-native";
-import Text from "../text/text";
+import Text from "../../text/text";
 import styles from "./button.styles";
 
 type ButtonProps = {
