--- src/components/button/button-gradient-RNA.tsx
@@ -1,28 +1,0 @@
-// import React, { ReactElement } from "react";
-// import { Button as RNEButton } from "react-native-elements";
-// import { LinearGradient, LinearGradientProps } from "react-native-linear-gradient";
-// import Text from "../text/text";
-// import styles from "./button.styles";
-
-// // type ButtonProps = {
-// //   title: string;
-// // } & LinearGradientProps;
-
-// export default function Button({ title, style, ...props }: any): ReactElement {
-//   return (
-//     <RNEButton
-//       // style={[styles.buttonContainer]}
-//       ViewComponent={LinearGradient}
-//       linearGradientProps={{
-//         colors: ["hsla(199, 79%, 71%, 1)", "hsla(220, 100%, 73%, 1)"],
-//         start: { x: 0, y: 1 },
-//         end: { x: 1, y: 1 }
-//         // style: { ...styles.ButtonGradientRadius }
-//       }}
-//     >
-//       <Text style={(styles.buttonText, { color: "#FFFFFF" })} weight="400">
-//         {title}
-//       </Text>
-//     </RNEButton>
-//   );
-// }
