+++ src/components/molecules/auth/submitButton.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
-import ButtonNormal from "../../button/button-normal";
-import ButtonGradient from "../../button/button-gradient";
+import ButtonNormal from "../../atoms/button/buttonNormal";
+import ButtonGradient from "../../atoms/button/buttonGradient";
 import { useFormikContext } from "formik";
 import style from "./submitButton.style";
